# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"LageplanResponse": {
				"properties": {
					"lageplan": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"EvaNumber": {
				"type": "string",
				"example": "8000105",
				"description": "Usually 7 digits, leading zeros can be omitted"
			},
			"TransportType": {
				"type": "string",
				"enum": [
					"HIGH_SPEED_TRAIN",
					"INTERCITY_TRAIN",
					"INTER_REGIONAL_TRAIN",
					"REGIONAL_TRAIN",
					"CITY_TRAIN",
					"SUBWAY",
					"TRAM",
					"BUS",
					"FERRY",
					"FLIGHT",
					"CAR",
					"TAXI",
					"SHUTTLE",
					"BIKE",
					"SCOOTER",
					"WALK",
					"UNKNOWN"
				],
				"description": "Type of transport. - HIGH_SPEED_TRAIN (High speed train [Hochgeschwindigkeitszug] like ICE or TGV etc.) - INTERCITY_TRAIN (Inter city train [Intercityzug]) - INTER_REGIONAL_TRAIN (Inter regional train [Interregiozug]) - REGIONAL_TRAIN (Regional train [Regionalzug]) - CITY_TRAIN (City train [S-Bahn]) - SUBWAY (Subway [U-Bahn]) - TRAM (Tram [Strassenbahn]) - BUS (Bus [Bus]) - FERRY (Ferry [Faehre]) - FLIGHT (Flight [Flugzeug]) - CAR (Car [Auto]) - TAXI (Taxi) - SHUTTLE (Shuttle [Ruftaxi]) - BIKE ((E-)Bike [Fahrrad]) - SCOOTER ((E-)Scooter [Roller]) - WALK (Walk ([Laufen]) - UNKNOWN (Unknown)"
			},
			"Coordinate2D": {
				"description": "2D coordinate within geo reference system.",
				"properties": {
					"latitude": {
						"type": "number",
						"format": "double",
						"description": "Latitude position in reference system."
					},
					"longitude": {
						"type": "number",
						"format": "double",
						"description": "Longitude position in reference system."
					}
				},
				"required": [
					"latitude",
					"longitude"
				],
				"type": "object",
				"additionalProperties": true
			},
			"GroupedStopPlace": {
				"properties": {
					"evaNumber": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"availableTransports": {
						"items": {
							"$ref": "#/components/schemas/TransportType"
						},
						"type": "array"
					},
					"position": {
						"$ref": "#/components/schemas/Coordinate2D"
					},
					"ifopt": {
						"type": "string"
					},
					"ril100": {
						"type": "string"
					},
					"alternativeRil100": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"stationId": {
						"type": "string"
					},
					"uic": {
						"type": "string"
					}
				},
				"required": [
					"evaNumber",
					"name",
					"availableTransports"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Record_string.RouteAuslastung-or-null_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"TrainOccupancyList": {
				"$ref": "#/components/schemas/Record_string.RouteAuslastung-or-null_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "Bahn.expert",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/stopPlace/v1/lageplan/{stopPlaceName}/{evaNumber}": {
			"get": {
				"operationId": "Lageplan",
				"responses": {
					"200": {
						"description": "URL to DB, HVV or NAHSH Lageplan",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LageplanResponse"
								}
							}
						}
					}
				},
				"tags": [
					"StopPlace"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "stopPlaceName",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "evaNumber",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/EvaNumber"
						}
					}
				]
			}
		},
		"/stopPlace/v1/search/{searchTerm}": {
			"get": {
				"operationId": "StopPlaceSearch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GroupedStopPlace"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "This might fall back to use HAFAS",
				"tags": [
					"StopPlace"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "searchTerm",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "max",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					},
					{
						"description": "Only returns stopPlaces iris-tts can handle (/abfahrten)",
						"in": "query",
						"name": "filterForIris",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "groupedBySales",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				]
			}
		},
		"/stopPlace/v1/{evaNumberOrRl100}": {
			"get": {
				"operationId": "StopPlaceByEvaOrRl100",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GroupedStopPlace"
								}
							}
						}
					},
					"404": {
						"description": ""
					}
				},
				"tags": [
					"StopPlace"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "evaNumberOrRl100",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/stopPlace/v1/{evaNumber}/trainOccupancy": {
			"get": {
				"operationId": "TrainOccupancy",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TrainOccupancyList"
								}
							}
						}
					},
					"404": {
						"description": ""
					}
				},
				"description": "Currently only for VRR. <br>\nSource: https://github.com/derf/eva-to-efa-gw<br>\nThanks derf.",
				"tags": [
					"StopPlace"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "evaNumber",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/EvaNumber"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "https://bahn.expert/api"
		}
	],
	"tags": [
		{
			"name": "HAFAS",
			"description": "All Operation use a HAFAS backend. By default it will be the HAFAS backend by DB Vertrieb (DB Navigator). Different backends can be selected by providing the query paramter profile."
		},
		{
			"name": "Journeys"
		},
		{
			"name": "IRIS"
		},
		{
			"name": "CoachSequence"
		},
		{
			"name": "StopPlace"
		}
	]
}